+++ packages/MeetupCom/src/Guzzle/Oauth2AwareClient.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Fop\MeetupCom\Guzzle;
+
+use GuzzleHttp\Client;
+
+final class Oauth2AwareClient extends Client
+{
+}
